Add solicitors to defence requests
We will seed our solicitor users in the same way 
as the search api for alpha, so any solicitor
returned will correspond to a login in our system.

There is a label still being rendered for email
on the form but I am leaving it for now as the
form will change a lot with rebase.