Restrict editing of opened/accepted Drs
After a DR has been opened or accepted, no longer
allow editing of details that may be critical to
the integrity of the request. Ie, if the age of
the detainee changes then it may change the need
for a solicitor to be present, in these cases a 
request should be closed and recreated with
correct details.