Fix testing of Defence Request transitions
The existing transition testing code was left over
from using the `simple_states` gem, so it was
accidentally not testing anything in a lot of
places: eg. asserting that a statement doesn't
throw an exception instead of returning false on
failure, or asserting that a model cannot perform
a transition that does not exist.