Delegate form fields to those before casting
Before: filling in a field like Age as a string 
would cast the value to `0`, which would then be 
resurfaced when the page was re-rendered with
errors. Instead we can delegate to the attribute
before it was cast to refill the form with the
exact values that were input before the attempt
to submit.