Generate dscc number randomly
Instead of pulling out all the dscc numbers and
sorting them every time we want to create a new
one, we could keep trying randomly generated
values until one is valid.

We can't do this via callbacks on the dscc number, 
or as a transition event on the defence request.
So the best option is a new acknowledger class to
handle this behaviour.