Extracts Dashboard Resource
This extracts a dashboard resource from DefenceRequestsController:
this enables a simpler `DefenceRequestsController` and offers an object
that can take be extended without further overloading the views or the
defence request model.

It also gets eliminates unnecessary queries that were made by
"DefenceRequestsController#get_defence_request_scopes" and corresponding
duplicate calls to `render` in the view. The css classes passed down to
render as `type` were also seemingly only used to facilitate testing.
In a previous PR, it was ensured that each test creates only the set of
data it needs and, therefore, these have become unnecessary as well.