Convert dscc_number to text, and add trigram index
The purpose of the cast is because the index will
only work on text fields. The index will allow
searching via regular expressions to be much
faster (which will be part of the upcoming number
generation method)

Using gist opposed to gin index as the data will
be changing a lot, gist is 10 times faster to
index, but 3 times slower to recall.

We cannot create a unique gist index, but it
should not matter as we will be using a lock
to handle race conditions, and will validate
uniqueness in AR as an extra check.