Specialise defence request transitions API
This allows each defence request transition strategy to have a hard
dependency on knowing how to call both the transition predicate method
and the transition method (`.can_transition_type?` and `.transition_type`)
on the passed in defence request.

This is justifiable to the extend that strategies are specialized around
a specific transition and knowledge of methods calls is the lowest form
of coupling possible.

The trade-off implied allows us to remove the use of exceptions for
control flow and passing down the `transition_to` key continuously,
along with a more cohesive, specialised API for each strategy.