Persist registry services in separate store class
In development mode, rails likes to reload
classes. Unfortunately this means our registry
gets clobbered, and a new singleton object is
instantiated - losing any services we registered
in initialisers.

Keeping a separate class to use as the store
protects the service hash itself from being
recreated (due to rails only expecting the
ServiceRegistry class to be residing in that
file)